Avoid traversing irrelevant config in `substituteInTerm` (#127)
* Avoid traversing irrelevant config in `substituteInTerm`

`substituteInTerm` was traversing all parts of the term that contains variables, regardless of what it was actually substituting. The change avoids traversal when the contained variables are not the substituted ones. This matters for large configurations like the KEVM one, but did not show in the IMP examples.

Had to use an ignore pragma for hlint because of [an hlint bug with `OverloadedRecordDot`](https://github.com/ndmitchell/hlint/issues/1458)

* switch off unrecognised-pragma warning