Compute subsort closure for all partial definitions when internalising
Previously, complete subsort information (transitive subsorts) were
only available for the main module, and computed at the end of the
internalisation.  This change is to compute correct subsort
information for all definitions held in the `definitionMap` during
internalisation, so we can retain this map and use it to add more
modules.