Draft code to use symbolic predicate simplification
This reveals problems with the current setup of the code.

We need to simplify/evaluate (both!) a boolean term while inside
another equation application: The "mode" of function evaluation or
simplification cannot be a monad parameter

Also, the monad should not be nested (otherwise we could of course),
so we can detect if the execution loops

Furthermore, the original description suggests to use function
evaluation _and then_ simplification (if nothing has changed), and
iterate the two together in a loop.

The outer functions need to be rewritten, and the state must not by
itself determine what kind of equation we are applying (we want both).