Apply function equations and simplifications in each iteration
As described in #108, the equations should be applied in an alternating fashion.
The original description prefers functions but this code parameterises that preference.

Functions `simplify` and `evaluate` (formerly `evaluateFunctions`) are now exported for testing only and will in fact apply _both_ kinds of equations, so the theories for unit-testing focus on just one kind each. All tests still pass unmodified.