Return indeterminate result (not failed) when matching unevaluated function calls
Fixes #212

We were prematurely returning `MatchFailed` when domain values or
injections in the pattern did not correspond to the subject.

If the subject contains an unevaluated function call, the result needs
to be `MatchIndeterminate` instead, preventing the evaluator from
applying rules at lower priority.

An integration test with a minimal example (from #212) will be added, too.