Return `vacuous` when all applicable rules rewrite to bottom (#272)
Fixes #268

This PR introduces a new `RewriteRuleAppT` monad which is analogous to
the previously used `MaybeT`, but instead of returning only one kind of
failure, namely `Nothing`, we can instead distinguish a failure with
`NotApplicable` or `Trivial`. When we collect the results from
attempting rule application, we return a `RewriteTrivial` if all the
applied rules that we attempted resulted in the ensures clause of the
rule evaluating to bottom.