Fix: return simplified next states (#307)
It turns out, we were not returning the simplified states, but rather
the ones before simplification. Now here comes the fun part:
* returning `filteredNexts` in `Branching`, as originally intended,
breaks an integration tests, specifically
https://github.com/runtimeverification/hs-backend-booster/blob/main/test/rpc-integration/test-no-evaluator/
* The test brakes because `#Not (#Exists X. X #Equals 12)` is simplified
into `#Not (X #Equals 12)` (provided I understand the kore json
corectly).

The original response we store for this test is returned by
`kore-rpc-dev` as well, hence I make the conclusion that the
simplification that strips the `#Exists` happens in Booster.

I have to admit that the behavior of the original test is unclear to me.
We want to test that we do not perform LLVM simplifications on a symbol
with `no-evaluators`. That seems to indeed be that case. However, I do
not understand what Kore does once Booster aborts: we get the branching
result in the end.

**Update**: after discussing it, we've decided to only run
test-no-evaluator with `booster-dev`, since running this test in
`kore-rpc` (and thus `kore-rpc-booster`) triggers a counter-intuitive,
but correct, behavior. We don't want to test that one. I've regenerated
the response for it and removed the old one.