Siplmify and substitute unification remainders when applying rules (#333)
This PR modifies the `performRewrite` function in the following way:
* if any of the rule applications failed with `RuleApplicationUnclear`
(unification is unclear), we get the unification remainders, attempt to
simplify and substitute them into the term, and then retry applying
rules
* mild refactoring to make the state in `performRewrite` a little more
manageable

The diagram preceeding `perfromRewrite` needs to be updated.
@goodlyrottenapple could you have a look at the code and also let me
know which tool you used to produce the diagram?

---------

Co-authored-by: Samuel Balco <goodlyrottenapple@gmail.com>