431 handle raw term from llvm (#438)
Fixes #431 

During deserialisation of terms received from the LLVM backend,
transform `rawTerm(inj{S, SortKItem}(X))` into a plain `X` (of expected
sort `S`, checked during the transformation).

This code is added so that we can handle both the previous behaviour (of
adding an injection `inj{SortX , SortKItem}(X)` to all simplified terms
`X:SortX`) and the new behaviour of decorating the former with a
`rawTerm(_)`.

A follow-up PR will remove the code that previously handled the
`SortKItem` injection.