HOTFIX: restore subsort injections removed by LLVM
Terms returned from the LLVM simplification are wrapped into a
top-level `rawTerm(inj{S, SortKItem{}}(_))` which is removed
when deserialising the result term from the returned string.

However, when the returned term is of a subsort A of the required
sort B, there would be a chain of injections `A -> B -> KItem` which
is shortened out to `A -> KItem` by the LLVM simplifier, apparently.

In `simplifyTerm` we restore the required `A -> B` injection when it
is necessary, checking that `A` is indeed a subsort of `B` beforehand.
If the result sort `A` is not a subsort of the required sort `B`, an
error message is printed and the original unsimplified term is
returned.