Stronger map matching
Previously, only fully concrete maps (all keys and values identical)
would have matched, which is highly unlikely when matching
simplification rules.

Now we compare the key sets for (syntactic) equality after
substituting with the current substitution, and then try to perform
matching on the remainder if a solution is at all possible.

Most importantly, maps of fully-concrete keys can now be handled,
matching the contained values. In addition, the case of matching a
single unknown key out of a singleton map (i.e., without opaque rest
in the subject map) is now possible.

Any ambiguous case is still rejected as `MatchIndeterminate` but some
cases can be decided as `MatchFailed` immediately.