Add conservative hooks for MAP.{lookup,lookupOrDefault,in_keys}
As an alternative to modifying rules in `domains.md`, this change
implements 3 crucial hooks for internalised `Map`s:

* `lookup`: partial function that looks up a value for a given key in
  a map. Returns no result when the key is not present.
* `lookupOrDefault`: total function that looks up a value for a given
  key in a map. If the key is not present, it returns a given default
  value instead. The implemented hook won't return any result when the
  presence of the key cannot be decided.
* `in_keys`: predicate that checks whether a given key is present in a
  map. The implemented hook won't return any result if the presence of
  the key cannot be decided.

All three hooks are conservative, not returning any results when they
find unevaluated (parts of the) arguments.

These hooks, together with the enhanced `Map`-matching of #502, avoid
a fall-back to legacy `kore-rpc` in an issue reported in #498.

Alternatively, more suitable rules for `in_keys` could be added to
`domains.md`, and the rules for `lookup` and `lookupOrDefault` could
be marked as `preserves-definedness` so `booster` will use them.